app: Add support for Android usage and last run time (API 21+).
This commit adds optional support for Android usage. This requires a system
permission, something we will never ask for, but we give a handy link in the
preferences screen.

The Android usage goes by time rather than quantity, this commit adds a field
for time storage in the LaunchableActivity.

This commit also preserves all old behaviour, so the permission can be
switched on/off easily without affecting normal behaviour.

With Android usage enabled, the order will be much more accurate, in any case
with exception of any inaccuracies Android might add (to prevent abuse).

Finally, this will not be available on all versions of Android, notably
your milage might vary depending on Android vendor.