SearchActivity: Add navigation bar tweaks.
This commit is not optimal, but a requirement due to the way Android
alternative ROMs allow navigation bar modifications and the lack of an SDK
for these changes.

If Android's SDK wasn't a garbage can fire, there would be some kind of SDK to
be able to accurately tell WHERE the navigation bar is, and whether the
navigation bar is translucent or not. The defaults work should work in
standard roms, these tweaks help in edge cases.

I hope to be able to revert this commit one day, but until then, the
tweaks are necessary for these corner cases.