SearchActivity: Workaround 180 degree landscape orientation change issue.
Prior to this commit, 180 degree landscape orientation change did not modify
the padding after the orientation change. This is due to there being no
Activity lifecycle callbacks being performed on 180 degree orientation change.
This is probably based off an assumption made in the design of Android that
nothing will need to change on 180 degree orientation change.

This change uses DisplayManager/DisplayListener to detect orientation changes.
I found OrientationEventListener to be unsuitable for this implementation as
small orders of magnitude are not detected, and this ensures that a callback
occurs /every/ time. It also requires far fewer calculations.