SearchActivity: Cache on label load.
The one thing that has stuck in my craw during the entire time I've looked at
this launcher is the (very) slow load time. In the past we've worked around
this problem by heavily threading the label load (via
LoadLaunchableActivityTask). This commit removes the threading and replaces it
with a SharedPreferences cache, which is more performant and doesn't
add the unnecessary complexity associated with LoadLaunchableActivityTask.