KeikaiLauncher: Perform the remainder of the name changes.
This commit removes the Proguard stuff, it's outdated, I'm not sure how to
work with Proguard and I don't really care. PRs welcome.

This also iterates the version to "2.0.0" and the version code. This iteration
is not a sign of stability, but differentiation from HayaiLauncher. Long live
HayaiLauncher.

Thanks to the users from Issue #6 for kicking me hard enough to finally make
this happen. <3