app: Support work profiles, change app loader for API 24 and higher.
This commit adds support for work profiles, which in turn depends on a newer
app loader, LauncherApps which requires less work for this app to do, which
should result in a quicker inital load (this commit will resolve
KeikaiLauncher/KeikaiLauncher#15).

Set to API 24 rather than API 21 due to inconsistencies in the APIs between 21 and 24.