app: Refactor layout.
This commit removes all the crufty tweaks and hacks from the layout. I'm
interested in hearing about bugs, but I'm not going to regress on inserting
hacks to make things work anymore. They caused the layout to get too fragile
on various API levels and were unmanagable.

This also changes the visuals to allow the icons to scroll "underneath" things.